Code review follow-up: namespace fix, PositionTolerance as double, scoped TLS bypass
- Move MarkerPoint/ZoneMatch to VisualMountParking.Markers namespace
- Change Config.PositionTolerance from decimal to double, keep the
  decimal conversion confined to SettingsForm's NumericUpDown
- Replace the process-wide ServicePointManager certificate bypass with
  a per-HttpClient ServerCertificateCustomValidationCallback, scoped to
  the two HttpClient instances that talk to the Reolink camera

+++ VisualMountParking/AutoPark.cs
@@ -121,7 +121,7 @@ namespace VisualMountParking
                 }
                 else
                 {
-                    var maxdelta = (double) _Config.PositionTolerance;
+                    var maxdelta = _Config.PositionTolerance;
                     InRange = ar.GetDistance() <= maxdelta && dec.GetDistance() < maxdelta;
                 }
             });
